requirements.txt
# Dependencias del sistema POS AFIP
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
zeep==4.2.1
cryptography==41.0.4
requests==2.31.0
python-dateutil==2.8.2
lxml==4.9.3
urllib3==2.0.4
